pyproject.toml
[project]
name = "langlearner"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Yuya Fujiwara", email = "y.fujiwara@petapa.jp" }
]
dependencies = [
    "streamlit>=1.29.0",
    "langchain>=0.0.27",
    "openai>=1.6.1",
    "st-multimodal-chatinput>=0.1.6",
    "google-generativeai>=0.3.1",
    "langchain-core>=0.1.1",
    "pvporcupine>=3.0.1",
    "google-cloud-speech>=2.23.0",
    "google-cloud-texttospeech>=2.15.0",
    "PyAudio>=0.2.14",
    "opencv-python>=4.8.1.78",
    "pydub>=0.25.1",
    "Pillow>=10.1.0",
    "python-dotenv>=1.0.0",
    "torch>=2.1.2",
    "torchvision>=0.16.2",
    "torchaudio>=2.1.2",
    "py3-tts>=3.5",
    "blobfile>=2.1.1",
    "Wave>=0.0.2",
    "openai-whisper>=20231117",
    "PyYAML>=6.0.1",
    "pygame>=2.5.2",
    "soundfile>=0.12.1",
    "pyttsx3>=2.90",
    "setuptools>=69.0.3",
    "sounddevice>=0.4.6",
    "numpy>=1.24.4",
    "transformers>=4.36.2",
    "accelerate>=0.25.0",
    "langchain_google_genai>=0.0.5",
    "ipykernel>=6.28.0",
    "wavefile>=1.6.2",
    "pysndfx>=0.3.6",
    "librosa>=0.10.1",
]
readme = "README.md"
requires-python = ">= 3.9"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src"]
